doc/requirements.txt
gorilla
sphinxcontrib-napoleon
